+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - Add `Deployer.send_with_receipt -> Result<(Contract, Receipt), Error>`
   so that the receipt can be returned to the called when deploying
   a contract [#865](https://github.com/gakonst/ethers-rs/pull/865)
+- Add Arbitrum mainnet and testnet to the list of known chains
 
 ## ethers-contract-abigen